requirements.txt
flask>=3.0.0
openai>=1.43.0,<2.0.0
python-dotenv>=1.0.1
requests>=2.32.0
